feat: 카카오 로그인 추가

+++ src/main/java/DMU/demo/chat/dto/ChatDto.java
@@ -8,7 +8,7 @@ import java.sql.Timestamp;
 @Data
 @Builder
 public class ChatDto {
-    private int sender;
+    private long sender;
     private String message;
     private Timestamp messageDate;
     private Time messageTime;
